Consolidate ForthControlWord and ForthMultiLine
Removes ForthControlWord and moves it's functionality into ForthMultiLine, with a variable @no_obj
which enables ForthMultiLine's original behavior when true, and behaves like ForthControlWord when false.
Simplifies ForthIf as well to read until a 'then', then split on 'else' if it exists, instead of calling
read_until again if an 'else' is found.