Remove now-useless stop_if_empty
Previously, all keywords would take a stop_if_empty variable, that would determine
if they should stop parsing if the end of the line was reached, or keep reading from the input source.
This was so that synatax checking could be done inside word definitions, as inside a word def everything
was on a single line, and reaching the end of the line meant a missing terminiating word. Now, since
objects are created before being stored in the word definition, the conditions for stop_if_empty to apply
never occur.