Move ForthKeyWord and subclasses into module, for proper namespacing.
ForthKeyWord class is renamed to ForthOp, and it and all subclasses are now part of the ForthOps module. This fixes the namespace
lookup with Module.const_get() which is used to convert input strings to ForthOps classes, as now the lookup is properly restricted to
only classes within the ForthOps namespace. (Previously, typing 'interpreter' would cause the program to crash because it tried to
create a ForthInterpreter class despite this not being a subclass of the ForthKeyWord/ForthOp class)